requirements.txt
Django==3.2
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.6.0.66
Pillow==9.5.0
torch==1.10.2
torchvision==0.11.3
